+++ src/pages/shared/users/index.js
@@ -1,7 +1,9 @@
 import callback from './callback'
 import manage from './manage'
+import groupEdit from './group_edit'
 
 export default {
   callback,
-  manage
+  manage,
+  groupEdit
 }